# Feature/remove spaced events