--- a/doc/README.md
+++ b/doc/README.md
@@ -126,25 +126,3 @@ el.trigger('start', { fuel: 89 }, true)
 ```
 
 @returns `el`
-
-### <a name="namespacing"></a> namespacing
-
-Events can be namespaced on a single level by using a `.` delimiter. Namespaced events listen to the primary event and can also be specifically triggered or removed.
-
-``` js
-// listen to start and start.honda events
-el.on('start.honda', function() {
-})
-
-// trigger all start events (including start.honda)
-el.trigger('start')
-
-// trigger only start.honda events
-el.trigger('start.honda')
-
-// remove only honda start events
-el.off('start.honda')
-
-// remove all start events (including start.honda)
-el.off('start')
-```