--- a/lib/index.js
+++ b/lib/index.js
@@ -18,17 +18,13 @@ var observable = function(el) {
    */
 
   /**
-   * Helper function needed to get and loop all the events in a string
-   * @param   { String }   e - event string
+   * Helper function needed to get and loop all the event in a string
+   * @param   { String }   name - event string
    * @param   {Function}   fn - callback
    */
-  function onEachEvent(e, fn) {
-    var es = e.split(' '), l = es.length, i = 0, name, indx
-    for (; i < l; i++) {
-      name = es[i]
-      indx = name.indexOf('.')
-      if (name) fn( ~indx ? name.substring(0, indx) : name, i, ~indx ? name.slice(indx + 1) : null)
-    }
+  function parseEvent(name, fn) {
+    var indx = name.indexOf('.')
+    if (name) fn(~indx ? name.substring(0, indx) : name, ~indx ? name.slice(indx + 1) : null)
   }
 
   /**
@@ -38,19 +34,18 @@ var observable = function(el) {
   // extend the el object adding the observable methods
   Object.defineProperties(el, {
     /**
-     * Listen to the given space separated list of `events` and
+     * Listen to the given `event` ands
      * execute the `callback` each time an event is triggered.
-     * @param  { String } events - events ids
+     * @param  { String } event - event id
      * @param  { Function } fn - callback function
      * @returns { Object } el
      */
     on: {
-      value: function(events, fn) {
+      value: function(event, fn) {
         if (typeof fn != 'function')  return el
 
-        onEachEvent(events, function(name, pos, ns) {
+        parseEvent(event, function(name, ns) {
           (callbacks[name] = callbacks[name] || []).push(fn)
-          fn.typed = pos > 0
           fn.ns = ns
         })
 
@@ -62,16 +57,16 @@ var observable = function(el) {
     },
 
     /**
-     * Removes the given space separated list of `events` listeners
-     * @param   { String } events - events ids
+     * Removes the given `event` listeners
+     * @param   { String } event - event id
      * @param   { Function } fn - callback function
      * @returns { Object } el
      */
     off: {
-      value: function(events, fn) {
-        if (events == '*' && !fn) callbacks = {}
+      value: function(event, fn) {
+        if (event == '*' && !fn) callbacks = {}
         else {
-          onEachEvent(events, function(name, pos, ns) {
+          parseEvent(event, function(name, ns) {
             if (fn || ns) {
               var arr = callbacks[name]
               for (var i = 0, cb; cb = arr && arr[i]; ++i) {
@@ -88,19 +83,19 @@ var observable = function(el) {
     },
 
     /**
-     * Listen to the given space separated list of `events` and
+     * Listen to the given `event` and
      * execute the `callback` at most once
-     * @param   { String } events - events ids
+     * @param   { String } event - event id
      * @param   { Function } fn - callback function
      * @returns { Object } el
      */
     one: {
-      value: function(events, fn) {
+      value: function(event, fn) {
         function on() {
-          el.off(events, on)
+          el.off(event, on)
           fn.apply(el, arguments)
         }
-        return el.on(events, on)
+        return el.on(event, on)
       },
       enumerable: false,
       writable: false,
@@ -109,12 +104,12 @@ var observable = function(el) {
 
     /**
      * Execute all callback functions that listen to
-     * the given space separated list of `events`
-     * @param   { String } events - events ids
+     * the given `event`
+     * @param   { String } event - event id
      * @returns { Object } el
      */
     trigger: {
-      value: function(events) {
+      value: function(event) {
 
         // getting the arguments
         var arglen = arguments.length - 1,
@@ -125,12 +120,12 @@ var observable = function(el) {
           args[i] = arguments[i + 1] // skip first argument
         }
 
-        onEachEvent(events, function(name, pos, ns) {
+        parseEvent(event, function(name, ns) {
 
           fns = slice.call(callbacks[name] || [], 0)
 
           for (var i = 0, fn; fn = fns[i]; ++i) {
-            if (!ns || fn.ns == ns) fn.apply(el, fn.typed ? [name].concat(args) : args)
+            if (!ns || fn.ns == ns) fn.apply(el, args)
             if (fns[i] !== fn) { i-- }
           }