--- a/package.json
+++ b/package.json
@@ -21,15 +21,15 @@
     "events"
   ],
   "devDependencies": {
-    "benchmark": "^2.1.0",
-    "coveralls": "^2.11.9",
-    "eslint": "^2.8.0",
+    "benchmark": "^2.1.1",
+    "coveralls": "^2.11.11",
+    "eslint": "^3.1.1",
     "expect.js": "^0.3.1",
-    "karma": "^0.13.22",
-    "karma-coverage": "^0.5.5",
-    "karma-mocha": "^0.2.2",
-    "karma-phantomjs-launcher": "^1.0.0",
-    "mocha": "^2.4.5",
+    "karma": "^1.1.1",
+    "karma-coverage": "^1.1.1",
+    "karma-mocha": "^1.1.1",
+    "karma-phantomjs-launcher": "^1.0.1",
+    "mocha": "^2.5.3",
     "phantomjs-prebuilt": "^2.1.7"
   },
   "author": "Muut, Inc. and other contributors",