--- a/test/bench/speed.js
+++ b/test/bench/speed.js
@@ -39,21 +39,6 @@ suite
       el.off('*')
     }
   })
-  .add('new-observable#multiple events', function() {
-    iterationsCounter++
-    el.trigger('foo')
-  }, {
-    onStart() {
-      iterationsCounter = 0
-      eventsCounter = 0
-      var fn = () => eventsCounter++
-      el.on('foo bar', fn)
-    },
-    onComplete() {
-      expect(eventsCounter).to.be(iterationsCounter)
-      el.off('*')
-    }
-  })
   .add('old-observable#simple event', function() {
     iterationsCounter++
     oldEl.trigger('foo')
@@ -63,21 +48,6 @@ suite
       eventsCounter = 0
       var fn = () => eventsCounter++
       oldEl.on('foo', fn)
-    },
-    onComplete() {
-      expect(eventsCounter).to.be(iterationsCounter)
-      oldEl.off('*')
-    }
-  })
-  .add('old-observable#multiple events', function() {
-    iterationsCounter++
-    oldEl.trigger('foo')
-  }, {
-    onStart() {
-      iterationsCounter = 0
-      eventsCounter = 0
-      var fn = () => eventsCounter++
-      oldEl.on('foo bar', fn)
     },
     onComplete() {
       expect(eventsCounter).to.be(iterationsCounter)