This patch makes riot-observable ~22% faster and a lot simpler but it removes the spaced events support
I think it's totally worth if we want make riot@3.0.0 faster
Patched observable
```
new-observable#simple event x 2,847,010 ops/sec ±1.12% (91 runs sampled)
new-observable#namespaced event x 2,891,754 ops/sec ±1.09% (91 runs sampled)
```
Current observable
```
new-observable#simple event x 2,233,131 ops/sec ±1.65% (90 runs sampled)
new-observable#namespaced event x 2,336,702 ops/sec ±0.87% (93 runs sampled)
```
Please let me know what you think
@tipiirai @rsbondi @aMarCruz @cognitom @rogueg