@rvion 2016-07-23
I personally don't need them, I have no problem having them removed, and will be quite happy to see riot being faster