@5angel 2016-07-23
I agree, even when I can use spaced events, I usually want to pass separate sets of data, so I divide them anyway.