@juodumas 2016-07-24
I fully agree about removing this edge case -> making events simpler, but only because I haven't got anything using riot2.

But one more thing: does riot really need to support namespaced events, which you have recently pulled in for riot2 and riot3? Looks like exchanging one edge case over aother..