@GianlucaGuarini 2016-07-24
I guess the namespaced events must be removed because they are causing problems in our riot router https://github.com/riot/route/issues/63 I will update this pull request