@GianlucaGuarini 2016-07-24
Bench without namespaced events:
```
new-observable#simple event x 3,973,287 ops/sec ±2.02% (79 runs sampled)
```
I am going to merge this looking forward to testing it in riot@3.0.0