@juodumas 2016-07-25
Looks like riot-3 is all about cutting corner cases. I like that.