@Mouvedia 2016-08-03
Concretely this means that for `on`

    element.on('start stop', function(type) {});
should be written

    element.on('start', function() {});
    element.on('stop', function() {});
from now on?

Doesn't that warrant something like

    element
       .stop(function() {})
       .start(function() {});
?

That would be much more elegant.