@GianlucaGuarini 2016-08-07
@Mouvedia your api proposal does not really make sense here since you could already do it http://plnkr.co/edit/fpRNmDWxv7BorDprebdI?p=preview  even if I would not recommend it