@senica 2016-09-05
I was sad to see namespaced events go away. I rely heavily on global events happening to keep the individual components updated. Namespaces make it easy to do `app.on('somethingHappened.'+_riot_id)` and then in the 'before-unmount' event, turn it off: `app.off('somethingHappened.'+_riot_id)`. Not having namespaces, makes this extremely difficult.