@juodumas 2016-09-05
@senica: just create a mixin which calls `.on` immediately and calls `.off` in `before-unmount`.

I use `this.onoff(app, 'eventName', () => { do_something })`.