@GianlucaGuarini 2016-09-05
@senica relying on `_riot_id` to handle your app logic it's a symptom of code smell. `_riot_id` is a "private" internal property, it could change at any time and it does not belong to the riot api. For all those that want to still using the slow namespaced events shooting on their feets you can use this simple mixin http://plnkr.co/edit/N0AIQKyp5ZyagemgwZfi?p=preview