@senica 2016-09-06
@GianlucaGuarini Thanks for pointing out that _riot_id is internal. I realize this is not the place for this, but I would argue that an ID should be just that,  an identifier — regardless if meant to be internal or not. Saying that it "could" change at anytime and be disassociated with the tag does not classify it to be an identifier.

I must have confused myself at some point with how the observable api worked, because I had an issue where a self looping tag would not unbind the event, and my call stack just kept increasing. Perhaps, the issue was elsewhere and I didn't realize. In any case, after going back and testing and everything was working, nice work in removing the namespaces ;)